[KAFKA][YCQL] Common code base for message queue Sink #7584
Summary:
The common codebase for different message queues. This package is used by both Kafka Sink and Apache Pulsar Sink.
It defines AbstractSinkRecord, Mapping of record fields to Table column and table configs, etc.

I had to introduce a config parameter 'useEventTimeForWrite' to avoid failures in case event time is not supplied by the message queues. The failure was specific to YugabyteDB and not Cassandra.

The package is also compiled against the yugabytedb cql driver to avoid different versions of classes being loaded.

Test Plan: All the tests already included in the parent repos.

Reviewers: hbhanawat, nkumar, nikhil, zyu

Reviewed By: zyu

Differential Revision: https://phabricator.dev.yugabyte.com/D10865